+++ workoutlogServer/models/user.js
@@ -1,5 +1,5 @@
 module.exports = function(sequelize, DataTypes){
-    let User = sequelize.define('user', {
+     return sequelize.define('user', {
         username: DataTypes.STRING,
         passwordhash:DataTypes.STRING
     });
